+++ src/steps/account/index.ts
@@ -1,11 +1,19 @@
 import {
+  createDirectRelationship,
+  Entity,
   IntegrationStep,
   IntegrationStepExecutionContext,
+  RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
 import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
-import { Entities, Steps } from '../constants';
+import {
+  Entities,
+  Relationships,
+  SERVICE_ENTITY_KEY,
+  Steps,
+} from '../constants';
 import { createOrganizationEntity } from './converter';
 
 export async function fetchOrganizations({
@@ -13,9 +21,20 @@ export async function fetchOrganizations({
   jobState,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
   const apiClient = createAPIClient(instance.config);
+  const serviceEntity = (await jobState.getData(SERVICE_ENTITY_KEY)) as Entity;
 
   await apiClient.iterateOrganizations(async (orgData) => {
-    await jobState.addEntity(createOrganizationEntity(orgData));
+    const organizationEntity = await jobState.addEntity(
+      createOrganizationEntity(orgData),
+    );
+
+    await jobState.addRelationship(
+      createDirectRelationship({
+        _class: RelationshipClass.HAS,
+        to: serviceEntity,
+        from: organizationEntity,
+      }),
+    );
   });
 }
 
@@ -24,8 +43,8 @@ export const organizationSteps: IntegrationStep<IntegrationConfig>[] = [
     id: Steps.ORGANIZATIONS,
     name: 'Fetch Organization Details',
     entities: [Entities.ORGANIZATION],
-    relationships: [],
-    dependsOn: [],
+    relationships: [Relationships.ORGANIZATION_HAS_SERVICE],
+    dependsOn: [Steps.SERVICE],
     executionHandler: fetchOrganizations,
   },
 ];